implementar autenticacao jwt

+++ routes/tasks.js
@@ -2,12 +2,16 @@ module.exports = (app) => {
   const Tasks = app.db.models.Tasks;
   app
     .route("/tasks")
+    .all(app.auth.authenticate())
     .get((req, res) => {
-      Tasks.findAll({})
+      Tasks.findAll({
+        where: { user_id: req.user.id }
+      })
         .then((result) => res.json(result))
         .catch((error) => res.status(412).json({ msg: error.message }));
     })
     .post((req, res) => {
+      req.body.user_id = req.user.id;
       Tasks.create(req.body)
         .then((result) => res.json(result))
         .catch((error) => res.status(412).json({ msg: error.message }));
@@ -15,8 +19,9 @@ module.exports = (app) => {
 
   app
     .route("/tasks/:id")
+    .all(app.auth.authenticate())
     .get((req, res) => {
-      Tasks.findOne({ where: { id: req.params.id } })
+      Tasks.findOne({ where: { id: req.params.id, user_id: req.user.id } })
         .then((result) => {
           if (result) {
             res.json(result);
@@ -27,12 +32,12 @@ module.exports = (app) => {
         .catch((error) => res.status(412).json({ msg: error.message }));
     })
     .put((req, res) => {
-      Tasks.update(req.body, { where: { id: req.params.id } })
+      Tasks.update(req.body, { where: { id: req.params.id, user_id: req.user.id } })
         .then((result) => res.sendStatus(204))
         .catch((error) => res.status(412).json({ msg: error.message }));
     })
     .delete((req, res) => {
-      Tasks.destroy({ where: req.params })
+      Tasks.destroy({ where: req.params.id, user_id: req.user.id })
         .then((result) => res.sendStatus(204))
         .catch((error) => res.status(412).json({ msg: error.message }));
     });
